move files to welcome_page directory
