+++ .idea/misc.xml
@@ -28,6 +28,9 @@
         <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/bottom_sheet_verify_code.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/dop.xml" value="0.18697916666666667" />
         <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/test.xml" value="0.18697916666666667" />
+        <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/welcome_page.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/welcome_page/activity_main.xml" value="0.36102564102564105" />
+        <entry key="..\:/Users/shelk/AndroidStudioProjects/ChoiceItSamsungSchool/app/src/main/res/layout/welcome_page/bottom_sheet_forgot_password_activity.xml" value="0.36102564102564105" />
       </map>
     </option>
   </component>
