+++ app/src/main/java/com/example/choiceitsamsungschool/MainActivity.java
@@ -24,6 +24,7 @@ import android.widget.EditText;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.example.choiceitsamsungschool.welcome_page.TextInputLayoutTextWatcher;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.google.android.material.button.MaterialButton;
 import com.google.android.material.textfield.TextInputEditText;
